followed to 36:10 in video
