+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/BackgroundTask.java
@@ -1,8 +1,79 @@
 package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
+import android.util.Log;
+import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.util.FakeData;
+import org.jetbrains.annotations.NotNull;
+
 public abstract class BackgroundTask implements Runnable {
+
+    private static final String LOG_TAG = "BackgroundTask";
+
     public static final String SUCCESS_KEY = "success";
     public static final String USER_KEY = "user";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
+
+    public BackgroundTask(Handler messageHandler) {
+        this.messageHandler = messageHandler;
+    }
+
+    /**
+     * Message handler that will receive task results.
+     */
+    private Handler messageHandler;
+
+    protected void sendSuccessMessage() {
+        Bundle msgBundle = createBundle(true);
+        loadSuccessBundle(msgBundle);
+        sendMessage(msgBundle);
+    }
+
+    @Override
+    public void run() {
+        try {
+            runTask();
+            
+            sendSuccessMessage();
+        } catch (Exception ex) {
+            Log.e(LOG_TAG, ex.getMessage(), ex);
+            sendExceptionMessage(ex);
+        }
+    }
+
+    protected abstract void runTask();
+
+    protected abstract void loadSuccessBundle(Bundle msgBundle);
+
+    private void sendFailedMessage(String message) {
+        Bundle msgBundle = createBundle(false);
+        msgBundle.putString(MESSAGE_KEY, message);
+        sendMessage(msgBundle);
+    }
+
+    private void sendExceptionMessage(Exception exception) {
+        Bundle msgBundle = createBundle(false);
+        msgBundle.putSerializable(EXCEPTION_KEY, exception);
+        sendMessage(msgBundle);
+    }
+
+    private void sendMessage(Bundle msgBundle) {
+        Message msg = Message.obtain();
+        msg.setData(msgBundle);
+        messageHandler.sendMessage(msg);
+    }
+
+    @NotNull
+    private Bundle createBundle(boolean value) {
+        Bundle msgBundle = new Bundle();
+        msgBundle.putBoolean(SUCCESS_KEY, value);
+        return msgBundle;
+    }
+
+    protected FakeData getFakeData() {
+        return new FakeData();
+    }
 }
