+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowersTask.java
@@ -65,6 +65,8 @@ public class GetFollowersTask implements Runnable {
             List<User> followers = pageOfUsers.getFirst();
             boolean hasMorePages = pageOfUsers.getSecond();
 
+            //TODO: add runTask()
+
             sendSuccessMessage(followers, hasMorePages);
 
         } catch (Exception ex) {
