+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetUserTask.java
@@ -2,18 +2,15 @@ package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
-import android.os.Message;
-import android.util.Log;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
-import edu.byu.cs.tweeter.util.FakeData;
 
 /**
  * Background task that returns the profile for a specified user.
  */
 public class GetUserTask extends BackgroundTask {
-    private static final String LOG_TAG = "GetUserTask";
+//    private static final String LOG_TAG = "GetUserTask";
 
     /**
      * Auth token for logged-in user.
@@ -23,32 +20,15 @@ public class GetUserTask extends BackgroundTask {
      * Alias (or handle) for user whose profile is being retrieved.
      */
     private String alias;
-    /**
-     * Message handler that will receive task results.
-     */
-    private Handler messageHandler;
+//    /**
+//     * Message handler that will receive task results.
+//     */
+//    private Handler messageHandler;
 
     public GetUserTask(AuthToken authToken, String alias, Handler messageHandler) {
+        super(messageHandler);
         this.authToken = authToken;
         this.alias = alias;
-        this.messageHandler = messageHandler;
-    }
-
-    @Override
-    public void run() {
-        try {
-            User user = getUser();
-
-            sendSuccessMessage(user);
-
-        } catch (Exception ex) {
-            Log.e(LOG_TAG, ex.getMessage(), ex);
-            sendExceptionMessage(ex);
-        }
-    }
-
-    private FakeData getFakeData() {
-        return new FakeData();
     }
 
     private User getUser() {
@@ -56,36 +36,13 @@ public class GetUserTask extends BackgroundTask {
         return user;
     }
 
-    private void sendSuccessMessage(User user) {
-        Bundle msgBundle = new Bundle();
-        msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(USER_KEY, user);
-
-        Message msg = Message.obtain();
-        msg.setData(msgBundle);
-
-        messageHandler.sendMessage(msg);
+    @Override
+    protected void runTask() {
+        // TODO: This is empty only cuz we're using dummy data
     }
 
-    private void sendFailedMessage(String message) {
-        Bundle msgBundle = new Bundle();
-        msgBundle.putBoolean(SUCCESS_KEY, false);
-        msgBundle.putString(MESSAGE_KEY, message);
-
-        Message msg = Message.obtain();
-        msg.setData(msgBundle);
-
-        messageHandler.sendMessage(msg);
-    }
-
-    private void sendExceptionMessage(Exception exception) {
-        Bundle msgBundle = new Bundle();
-        msgBundle.putBoolean(SUCCESS_KEY, false);
-        msgBundle.putSerializable(EXCEPTION_KEY, exception);
-
-        Message msg = Message.obtain();
-        msg.setData(msgBundle);
-
-        messageHandler.sendMessage(msg);
+    @Override
+    protected void loadSuccessBundle(Bundle msgBundle) {
+        msgBundle.putSerializable(USER_KEY, getUser());
     }
 }
